requirements.txt
streamlit==1.31.1
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6